Abstract KFSimulator
